spike_research/requirements.txt
# Core FastAPI dependencies
fastapi==0.115.8
uvicorn==0.34.0
pydantic==2.10.6
python-multipart>=0.0.9

# Database dependencies
neo4j==5.28.1
pydruid[pandas,sqlalchemy]>=0.6.9
sqlalchemy>=1.4.0

# Data analysis and ML
pandas>=1.3.0
numpy
scipy

# Web and networking
requests==2.32.3
python-dotenv==1.0.1

# Optional: LLM integration (if needed for analysis)
openai==1.64.0

# Development tools
ruff>=0.1.0

# Core dependencies (auto-resolved)
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
h11==0.14.0
idna==3.10
pydantic_core==2.27.2
pytz==2025.1
sniffio==1.3.1
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0